aws: Add aws.DiscoverAddrsByEC2Tag func
This commit adds a new function, `aws.DiscoverAddrsByEC2Tag`, which
approximates the behaviour of HashiCorp's `go-discover` library (used in
Consul and Nomad). The function allows the IP addresses of all instances
with a tag matching a nominated value. The uses of this are far-
reaching, but primarily useful for finding the other members of an
autoscaling group in order to form a cluster.